src/openocd_mcp/requirements.txt
# Pinned dependencies for openocd-mcp service/runtime
pyinstaller==6.11.1
fastmcp==0.2.0
pyserial==3.5
pyelftools==0.31
cmsis-svd==0.5
pyusb==1.2.1
pytest==8.0.0
